size_t -> unsigned for matrix

+++ src/amun/common/base_matrix.cpp
@@ -5,16 +5,16 @@ using namespace std;
 
 namespace amunmt {
 
-size_t BaseMatrix::size() const {
-  size_t ret = dim(0);
-  for (size_t i = 1; i < SHAPE_SIZE; ++i) {
+unsigned BaseMatrix::size() const {
+  unsigned ret = dim(0);
+  for (unsigned i = 1; i < SHAPE_SIZE; ++i) {
     ret *= dim(i);
   }
 
   return ret;
 }
 
-std::string BaseMatrix::Debug(size_t detailed) const
+std::string BaseMatrix::Debug(unsigned detailed) const
 {
   std::stringstream strm;
   strm << dim(0) << "x" << dim(1) << "x" << dim(2) << "x" << dim(3) << "=" << size();
